+++ pkg/api/graphql/graphql.go
@@ -1,3 +1,5 @@
+//go:generate protoc -I. --gotemplate_out=all=true,template_dir=pkg/api/graphql/template:pkg/api/graphql --go_out=Minternal/api/domain/link/link.proto=.:. --go-grpc_out=Minternal/api/domain/link/link.proto=.:. --go_opt=paths=source_relative --go-grpc_opt=paths=source_relative internal/api/domain/link/link.proto
+
 package graphql
 
 import (
