+++ pkg/rpc/type.go
@@ -1,0 +1,9 @@
+package rpc
+
+import "google.golang.org/grpc"
+
+type RPCServer struct {
+	Run      func()
+	Server   *grpc.Server
+	Endpoint string
+}
